refactor into gateways and controllers
